# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Fragrance Rater",
    "description": "HTTP API for the Fragrance Rater service: a personal fragrance evaluation and recommendation system. Provides the fragrance catalog and an LLM-powered rating endpoint that scores a fragrance description and returns the model's reasoning. The rating endpoint calls OpenRouter (default model: `anthropic/claude-3.5-sonnet`); set `TEST_MODE=true` to use the bundled fixture response instead, which is what the Newman contract tests rely on. `POST /ratings` requires a shared household `X-API-Key` header and is rate limited; read-only endpoints do not require it.",
    "contact": {
      "name": "Byron Williams",
      "url": "https://github.com/ByronWilliamsCPA/fragrance-rater",
      "email": "byron@williamshome.family"
    },
    "license": {
      "name": "MIT"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/recommendation-measurement/runs": {
      "post": {
        "tags": [
          "recommendations"
        ],
        "summary": "Create Run",
        "description": "Generate recommendations and persist their impressions atomically.",
        "operationId": "create_run_api_v1_recommendation_measurement_runs_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RunCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/recommendation-measurement/runs/{run_id}": {
      "get": {
        "tags": [
          "recommendations"
        ],
        "summary": "Get Run",
        "description": "Return an existing run; repeated views create no new impressions.",
        "operationId": "get_run_api_v1_recommendation_measurement_runs__run_id__get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/recommendation-measurement/impressions/{impression_id}/responses": {
      "post": {
        "tags": [
          "recommendations"
        ],
        "summary": "Append Response",
        "description": "Append feedback, preserving every prior response revision.",
        "operationId": "append_response_api_v1_recommendation_measurement_impressions__impression_id__responses_post",
        "parameters": [
          {
            "name": "impression_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Impression Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResponseCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/recommendation-measurement/operational-events": {
      "post": {
        "tags": [
          "recommendations"
        ],
        "summary": "Operational Event",
        "description": "Record a pilot connectivity failure or subsequent manual recovery.",
        "operationId": "operational_event_api_v1_recommendation_measurement_operational_events_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OperationalEventCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Operational Event Api V1 Recommendation Measurement Operational Events Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ImpressionView": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "fragrance_id": {
            "type": "string",
            "title": "Fragrance Id"
          },
          "fragrance_name": {
            "type": "string",
            "title": "Fragrance Name"
          },
          "fragrance_brand": {
            "type": "string",
            "title": "Fragrance Brand"
          },
          "rank": {
            "type": "integer",
            "title": "Rank"
          },
          "score_type": {
            "type": "string",
            "title": "Score Type"
          },
          "score_value": {
            "type": "number",
            "title": "Score Value"
          },
          "match_percent": {
            "type": "integer",
            "title": "Match Percent"
          }
        },
        "type": "object",
        "required": [
          "id",
          "fragrance_id",
          "fragrance_name",
          "fragrance_brand",
          "rank",
          "score_type",
          "score_value",
          "match_percent"
        ],
        "title": "ImpressionView",
        "description": "Candidate and immutable ranking information shown to an evaluator."
      },
...
      "OperationalEventCreate": {
        "properties": {
          "reviewer_id": {
            "type": "string",
            "title": "Reviewer Id"
          },
          "event_type": {
            "type": "string",
            "enum": [
              "CONNECTIVITY_FAILURE",
              "MANUAL_RECOVERY"
            ],
            "title": "Event Type"
          },
          "details": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2000
              },
              {
                "type": "null"
              }
            ],
            "title": "Details"
          }
        },
        "type": "object",
        "required": [
          "reviewer_id",
          "event_type"
        ],
        "title": "OperationalEventCreate",
        "description": "Pilot connectivity or manual-recovery observation."
      },
...
      "ResponseCreate": {
        "properties": {
          "interested": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Interested"
          },
          "sampling_state": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "PLANNED",
                  "ACQUIRED",
                  "SAMPLED",
                  "UNAVAILABLE"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Sampling State"
          },
          "unavailable_reason": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2000
              },
              {
                "type": "null"
              }
            ],
            "title": "Unavailable Reason"
          },
          "outcome_evaluation_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Outcome Evaluation Id"
          },
          "outcome_observation_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Outcome Observation Id"
          },
          "would_wear": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Would Wear"
          },
          "would_buy": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Would Buy"
          }
        },
        "type": "object",
        "title": "ResponseCreate",
        "description": "Complete latest feedback state; submission appends a revision."
      },
...
      "ResponseView": {
        "properties": {
          "interested": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Interested"
          },
          "sampling_state": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "PLANNED",
                  "ACQUIRED",
                  "SAMPLED",
                  "UNAVAILABLE"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Sampling State"
          },
          "unavailable_reason": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2000
              },
              {
                "type": "null"
              }
            ],
            "title": "Unavailable Reason"
          },
          "outcome_evaluation_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Outcome Evaluation Id"
          },
          "outcome_observation_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Outcome Observation Id"
          },
          "would_wear": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Would Wear"
          },
          "would_buy": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Would Buy"
          },
          "id": {
            "type": "string",
            "title": "Id"
          },
          "impression_id": {
            "type": "string",
            "title": "Impression Id"
          },
          "revision": {
            "type": "integer",
            "title": "Revision"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "recorded_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Recorded By"
          }
        },
        "type": "object",
        "required": [
          "id",
          "impression_id",
          "revision",
          "created_at",
          "recorded_by"
        ],
        "title": "ResponseView",
        "description": "Stored append-only feedback revision."
      },
...
      "RunCreate": {
        "properties": {
          "reviewer_id": {
            "type": "string",
            "title": "Reviewer Id"
          },
          "limit": {
            "type": "integer",
            "maximum": 50.0,
            "minimum": 1.0,
            "title": "Limit",
            "default": 10
          },
          "exclude_rated": {
            "type": "boolean",
            "title": "Exclude Rated",
            "default": true
          },
          "candidate_strategy": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Candidate Strategy",
            "default": "catalog-affinity"
          }
        },
        "type": "object",
        "required": [
          "reviewer_id"
        ],
        "title": "RunCreate",
        "description": "Parameters that produce a new persisted recommendation run."
      },
      "RunView": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "reviewer_id": {
            "type": "string",
            "title": "Reviewer Id"
          },
          "algorithm_version": {
            "type": "string",
            "title": "Algorithm Version"
          },
          "candidate_strategy": {
            "type": "string",
            "title": "Candidate Strategy"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "impressions": {
            "items": {
              "$ref": "#/components/schemas/ImpressionView"
            },
            "type": "array",
            "title": "Impressions"
          }
        },
        "type": "object",
        "required": [
          "id",
          "reviewer_id",
          "algorithm_version",
          "candidate_strategy",
          "created_at",
          "impressions"
        ],
        "title": "RunView",
        "description": "A recommendation run and its already-persisted impressions."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}